user and task relation

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/user/user.entity';
 import { DeleteResult } from 'typeorm';
 import { v1 as uuid } from 'uuid';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -15,38 +16,43 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTask(filterDto);
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTask(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
-            throw new NotFoundException();
+            throw new NotFoundException(`Task with ID"${id}" not found`);
         }
 
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return await this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return await this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<DeleteResult> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<DeleteResult> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
+
+        if (result.affected === 0) {
+            throw new NotFoundException(`Task with ID"${id}" not found`);
+        }
+
         return result;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
 
-    async updateTask(id: number, title: string, description: string): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTask(id: number, title: string, description: string, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.title = title;
         task.description = description;
         await task.save();
